refactor(#3272): extract ast-modifiers + linear-type-reservations from index.ts (refs #3182)
Verbatim moves; index.ts re-exports for stable import paths. Byte-identity
IDENTICAL (39/39), tsc 0.

✓
